add session restore support
We did most of the work for this feature in a previous change
that refactored how we do a bunch of threading. This patch
takes advantage of that rewriting to actually add screen restore
support. By default, shpool still behaves the same way it always
has, but now it supports a new option, `session_restore_mode`,
which can either be `"simple"` (the default), `"screen"` (to restore
the entire screen, or `{ lines = n }` to restore the last n lines
on reconnect. Basically all the heavy lifting for this change
is done by the vt100 crate.

+++ src/daemon/config.rs
@@ -78,7 +78,7 @@ pub enum SessionRestoreMode {
     Screen,
     /// Emit enough output data to restore the last n lines of
     /// history from the output spool.
-    Lines(usize),
+    Lines(u16),
 }
 
 #[cfg(test)]
